refactor: Split utils into {time,pygame}_utils & move functions around
