+++ pucoti.py
@@ -1,3 +1,5 @@
+#!/usr/bin/env python
+
 from src.app import app
 
 if __name__ == "__main__":
