+++ src/constants.py
@@ -8,8 +8,7 @@ BELL = ASSETS / "bell.mp3"
 BIG_FONT = ASSETS / "Bevan-Regular.ttf"
 FONT = BIG_FONT
 WINDOW_SCALE = 1.2
-MIN_HEIGHT = 5
-MIN_WIDTH = 15
+MIN_WINDOW_SIZE = 15, 5
 POSITIONS = [(-5, -5), (5, 5), (5, -5), (-5, 5)]
 SHORTCUTS = """
 j k: -/+ 1 minute
@@ -31,6 +30,3 @@ PUCOTI
 """.strip()
 
 NUMBER_KEYS = [pg.K_0, pg.K_1, pg.K_2, pg.K_3, pg.K_4, pg.K_5, pg.K_6, pg.K_7, pg.K_8, pg.K_9]
-
-# Diegouses sway, and it needs a few tweaks as it's a non-standard window manager.
-RUNS_ON_SWAY = os.environ.get("SWAYSOCK") is not None
